moved last few common functions to utils
