+++ brainglobe_napari_io/brainreg/reader_dir_atlas_space.py
@@ -10,6 +10,7 @@ from napari.types import LayerDataTuple
 from brainglobe_napari_io.utils import (
     is_brainreg_dir,
     load_additional_downsampled_channels,
+    load_atlas,
 )
 
 PathOrPaths = Union[List[os.PathLike], os.PathLike]
@@ -82,25 +83,5 @@ def reader_function(path: os.PathLike) -> List[LayerDataTuple]:
         )
     )
     layers = load_atlas(atlas, layers)
-
-    return layers
-
-
-def load_atlas(
-    atlas: BrainGlobeAtlas, layers: List[LayerDataTuple]
-) -> List[LayerDataTuple]:
-    atlas_image = atlas.annotation
-    layers.append(
-        (
-            atlas_image,
-            {
-                "name": atlas.atlas_name,
-                "visible": False,
-                "blending": "additive",
-                "opacity": 0.3,
-            },
-            "labels",
-        )
-    )
 
     return layers
